--- backend/index.js
@@ -1,6 +1,0 @@
-import { app, dbSession } from "./src/app.js";
-
-app.listen(process.env.PORT || 3000, async () => {
-  await dbSession.setup();
-  console.log(`Listening to PORT: ${process.env.PORT || 3000}`);
-});
