+++ backend/src/utils.ts
@@ -1,0 +1,11 @@
+export function generateAlphaNumericString(length: number): string {
+	const characters =
+		"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+	let result = "";
+	for (let i = 0; i < length; i++) {
+		result += characters.charAt(
+			Math.floor(Math.random() * characters.length),
+		);
+	}
+	return result;
+}
